test: reorganise tests to match new structure
